Nav fix (#27)
* initial

* navbar mobile

* Bigger Screen Navbar and its Responsiveness

* improvised

* Responsiveness Fixed

* transition added in mobNav

* remove unnecessary marquees

* Fixed production errors

* Lockfile issues

* issues resolved

* minor change

* Delete .vscode directory

* Made necessaary changes except ALIAS

* improved navbar

* fixed build issue

---------
