+++ src/app/(main)/home/page.tsx
@@ -7,15 +7,15 @@ import About from "~/components/AboutPage";
 export default function Page() {
   return (
     <div className="bg-dotted flex min-h-screen w-full flex-col items-center">
-      <About/>
+      <About />
       <div className="px-10">
-        <PhotoGallery/>
+        <PhotoGallery />
       </div>
       <div className="text-gradient-blue mt-10 font-rp1 text-2xl font-bold md:hidden">
         Tecnoesis App
       </div>
-      <div className="flex w-full lg:py-[5rem] mb-10 flex-col gap-y-20">
-        <Download/>
+      <div className="mb-10 flex w-full flex-col gap-y-20 lg:py-[5rem]">
+        <Download />
       </div>
       <div className="text-gradient-blue m-10 font-rp1 text-2xl font-bold md:text-6xl">
         Past Sponsors
@@ -24,11 +24,11 @@ export default function Page() {
         <MarqueeComponent direction={"left"} set={1} />
         <MarqueeComponent direction={"right"} set={2} />
       </div>
-      <div className="text-gradient-blue mb-10 mt-20 font-rp1 text-2xl font-bold md:text-6xl sm:pt-20">
+      <div className="text-gradient-blue mb-10 mt-20 font-rp1 text-2xl font-bold sm:pt-20 md:text-6xl">
         Sponsors
       </div>
-      <div className="px-10 w-full">
-        <SponsorHome/>
+      <div className="w-full px-10">
+        <SponsorHome />
       </div>
     </div>
   );
