+++ src/app/(main)/layout.tsx
@@ -1,0 +1,12 @@
+import FinalNav from "~/components/Navbar/FinalNav";
+
+export default function RootLayout({
+  children,
+}: Readonly<{ children: React.ReactNode }>) {
+  return (
+    <div className="bg-dotted">
+      <FinalNav />
+      {children}
+    </div>
+  );
+}
