+++ src/app/layout.tsx
@@ -2,7 +2,6 @@ import "~/styles/globals.css";
 import { GeistSans } from "geist/font/sans";
 import { type Metadata } from "next";
 import { Toaster } from "sonner";
-import FinalNav from "~/components/Navbar/FinalNav";
 
 export const metadata: Metadata = {
   title: "Tecnoesis 2024",
@@ -26,9 +25,6 @@ export default function RootLayout({
   return (
     <html lang="en" className={`${GeistSans.variable}`}>
       <body>
-        {/* <div className="sticky top-0 w-screen z-20"> */}
-        <FinalNav />
-        {/* </div> */}
         {children}
         <Toaster
           toastOptions={toastOps}
